feat(axum): Modularize posts module and clean up imports
Authored-by: Gemini CLI
