+++ axum/src/models.rs
@@ -3,48 +3,6 @@ use serde::{Deserialize, Serialize};
 use sqlx::FromRow;
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
-pub struct Post {
-    pub id: i64,
-    pub title: String,
-    pub content: String,
-    pub created_at: DateTime<Utc>,
-    pub updated_at: DateTime<Utc>,
-    pub embedding: Option<pgvector::Vector>,
-    pub user_id: i64,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
-pub struct PostGraphData {
-    pub id: i64,
-    pub title: String,
-    pub embedding: Option<pgvector::Vector>,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct CreatePost {
-    pub title: String,
-    pub content: String,
-    //pub user_id: i64,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct UpdatePost {
-    pub title: Option<String>,
-    pub content: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PostResponse {
-    pub id: i64,
-    pub title: String,
-    pub content: String,
-    pub created_at: DateTime<Utc>,
-    pub updated_at: DateTime<Utc>,
-    pub user_id: i64,
-    pub related_posts: Vec<Post>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct User {
     pub id: i64,
     pub username: String,
@@ -74,23 +32,4 @@ pub struct CreateComment {
     pub content: String,
     pub post_id: i64,
     pub user_id: i64,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct GraphNode {
-    pub id: String,   // Post ID as string
-    pub name: String, // Post title
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct GraphLink {
-    pub source: String, // Source Post ID as string
-    pub target: String, // Target Post ID as string
-    pub value: f32,     // Similarity score
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct GraphData {
-    pub nodes: Vec<GraphNode>,
-    pub links: Vec<GraphLink>,
 }
