+++ axum/src/posts/handlers.rs
@@ -1,111 +1,25 @@
-// src/posts.rs
-
 use axum::{
-    Router,
     extract::{Json, Path, State},
     http::StatusCode,
-    middleware::{self, Next, from_fn},
-    routing::{delete, get, post, put},
 };
-
 use jwt_authorizer::JwtClaims;
-
-use crate::{
-    auth::UserClaims,
-    models::{CreatePost, Post, PostGraphData, PostResponse, UpdatePost, User},
-};
-use redis::{Client, aio::MultiplexedConnection};
+use pgvector::Vector;
 use serde::{Deserialize, Serialize};
 use sqlx::{Postgres, pool::Pool};
-
-//cache
-use axum_redis_cache::{CacheManager, CacheState};
-
-//embedding
-use pgvector::Vector;
-
-//graph
-use crate::models::{GraphData, GraphLink, GraphNode};
-
 use std::time::Duration;
 use tokio::time::sleep;
 
-pub fn routes(
-    //mut conn: MultiplexedConnection
-    cache_state: CacheState,
-) -> Router<Pool<Postgres>> {
-    Router::new()
-        .merge(post_routes_auth())
-        .merge(post_routes_cache().layer(middleware::from_fn_with_state(
-            cache_state,
-            axum_redis_cache::middleware,
-        )))
-}
-
-fn post_routes_auth() -> Router<Pool<Postgres>> {
-    Router::new()
-        .route("/posts", get(list_posts).post(create_post))
-        .route("/posts/graph", get(get_graph_data))
-}
-
-fn post_routes_cache() -> Router<Pool<Postgres>> {
-    Router::new().route(
-        "/posts/:id",
-        get(get_posts).delete(delete_post).put(update_post),
-    )
-}
-
-pub fn write_to_cache(old: String, new: String) -> String {
-    let parsed_body: UpdatePost = serde_json::from_str(&new).unwrap();
-    let mut payload: PostResponse = serde_json::from_str(&old).unwrap();
-
-    if let Some(content) = parsed_body.content {
-        payload.content = content;
-    }
-    if let Some(title) = parsed_body.title {
-        payload.title = title;
-    }
-    serde_json::to_string(&payload).unwrap()
-}
-
-pub async fn callback(db: Pool<Postgres>, value: String) {
-    use crate::models::PostResponse;
-    let json: PostResponse = serde_json::from_str(&value).unwrap();
-    use crate::models::UpdatePost;
-    use crate::posts::__update_post_from_cache;
-    //TODO : from PostResponse, to UpdatePost
-    let update_json = UpdatePost {
-        title: Some(json.title),
-        content: Some(json.content),
-    };
-    let _ = __update_post_from_cache(State(db), Path(json.id), Json(update_json)).await;
-}
-
-pub async fn delete_callback(db: Pool<Postgres>, key: String) {
-    if let Ok(post_id) = key.parse::<i64>() {
-        println!("🧹 expired 감지됨: delete marker for post_id={}", post_id);
-
-        // 실제 DB에서 삭제
-        let result = sqlx::query("DELETE FROM posts WHERE id = $1")
-            .bind(post_id)
-            .execute(&db)
-            .await
-            .unwrap();
-
-        println!(
-            "✅ DB에서 post {} 삭제 완료 ({} rows affected)",
-            post_id,
-            result.rows_affected()
-        );
-    }
-}
+use super::graph::get_related_post;
+use super::models::{CreatePost, Post, PostGraphData, PostResponse, UpdatePost};
+use super::utils::internal_error;
+use crate::auth::UserClaims;
 
 #[derive(Deserialize)]
-struct EmbedResponse {
-    embedding: Vec<f32>,
+pub struct EmbedResponse {
+    pub embedding: Vec<f32>,
 }
 
-async fn create_post(
+pub async fn create_post(
     State(db): State<Pool<Postgres>>,
     JwtClaims(user): JwtClaims<UserClaims>,
     Json(payload): Json<CreatePost>,
@@ -136,7 +50,7 @@ async fn create_post(
     Ok(Json(post))
 }
 
-async fn list_posts(
+pub async fn list_posts(
     State(db): State<Pool<Postgres>>,
     JwtClaims(user): JwtClaims<UserClaims>,
 ) -> Json<Vec<Post>> {
@@ -149,10 +63,11 @@ async fn list_posts(
     .await
     .unwrap();
 
-    //TODO : unwarp 사용 안하는게 좋음
+    //TODO : unwrap 사용 안하는게 좋음
     Json(posts)
 }
-async fn get_posts(
+
+pub async fn get_posts(
     State(db): State<Pool<Postgres>>,
     Path(id): Path<i64>,
     JwtClaims(user): JwtClaims<UserClaims>,
@@ -177,7 +92,8 @@ async fn get_posts(
 
     Ok(Json(post_response))
 }
-async fn delete_post(
+
+pub async fn delete_post(
     State(db): State<Pool<Postgres>>,
     Path(id): Path<i64>,
     JwtClaims(user): JwtClaims<UserClaims>,
@@ -192,7 +108,6 @@ async fn delete_post(
     Ok(())
 }
 
-//todo : make this
 pub async fn update_post(
     State(db): State<Pool<Postgres>>,
     JwtClaims(user): JwtClaims<UserClaims>,
@@ -308,99 +223,4 @@ pub async fn __update_post_from_cache(
     .unwrap();
 
     tx.commit().await.map_err(internal_error).unwrap();
-}
-
-async fn get_related_post(post: &Post, db: &Pool<Postgres>) -> Vec<Post> {
-    /* related post 가져오기기 */
-    let Some(embedding) = &post.embedding else {
-        return Vec::new();
-    };
-    // 유사도 기반으로 관련 포스트 3개 반환
-    let related_posts = sqlx::query_as::<_, Post>(
-        r#"
-        SELECT * FROM posts
-        WHERE user_id = $1
-        AND id != $2
-        AND embedding IS NOT NULL
-        ORDER BY embedding <-> $3
-        LIMIT 3
-        "#,
-    )
-    .bind(post.user_id)
-    .bind(post.id)
-    .bind(embedding)
-    .fetch_all(db)
-    .await
-    .unwrap_or_else(|e| {
-        eprintln!("DB error during related post fetch: {}", e);
-        Vec::new()
-    });
-
-    related_posts
-}
-
-fn internal_error<E: std::fmt::Display>(e: E) -> (StatusCode, String) {
-    (StatusCode::INTERNAL_SERVER_ERROR, format!("에러: {}", e))
-}
-
-async fn get_graph_data(
-    State(db): State<Pool<Postgres>>,
-    JwtClaims(user): JwtClaims<UserClaims>,
-) -> Result<Json<GraphData>, (StatusCode, String)> {
-    let posts = sqlx::query_as::<_, PostGraphData>(
-        r#"SELECT id, title, embedding
-            FROM posts
-            WHERE user_id = $1
-            AND embedding IS NOT NULL"#,
-    )
-    .bind(user.sub)
-    .fetch_all(&db)
-    .await
-    .map_err(internal_error)?;
-
-    let mut nodes: Vec<GraphNode> = Vec::new();
-    let mut links: Vec<GraphLink> = Vec::new();
-
-    for post in &posts {
-        nodes.push(GraphNode {
-            id: post.id.to_string(),
-            name: post.title.clone(),
-        });
-    }
-
-    // Calculate similarities and create links
-    for i in 0..posts.len() {
-        for j in (i + 1)..posts.len() {
-            let p1 = &posts[i];
-            let p2 = &posts[j];
-
-            if let (Some(e1), Some(e2)) = (&p1.embedding, &p2.embedding) {
-                // Calculate cosine distance (pgvector uses <-> for cosine distance)
-                //println!("Calculating distance between post {} and {}", p1.id, p2.id);
-                let distance = sqlx::query_scalar::<_, f64>("SELECT $1 <-> $2")
-                    .bind(e1)
-                    .bind(e2)
-                    .fetch_one(&db)
-                    .await
-                    .map_err(|e| {
-                        eprintln!("Error calculating distance: {}", e);
-                        internal_error(e)
-                    })?;
-                //println!("Distance: {}", distance);
-
-                // Only add link if similarity is above a certain threshold
-                // Cosine distance ranges from 0 to 2. 0 means identical, 2 means opposite.
-                // A distance of < 0.2 means similarity > 0.8 (since similarity = 1 - distance/2)
-                if distance < 0.5 {
-                    links.push(GraphLink {
-                        source: p1.id.to_string(),
-                        target: p2.id.to_string(),
-                        value: (1.0 - (distance / 2.0)) as f32, // Convert distance to similarity (0 to 1)
-                    });
-                }
-            }
-        }
-    }
-
-    Ok(Json(GraphData { nodes, links }))
 }
