+++ axum/src/posts/models.rs
@@ -1,0 +1,64 @@
+use chrono::{DateTime, Utc};
+use pgvector::Vector;
+use serde::{Deserialize, Serialize};
+use sqlx::FromRow;
+
+#[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
+pub struct Post {
+    pub id: i64,
+    pub title: String,
+    pub content: String,
+    pub created_at: DateTime<Utc>,
+    pub updated_at: DateTime<Utc>,
+    pub embedding: Option<pgvector::Vector>,
+    pub user_id: i64,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
+pub struct PostGraphData {
+    pub id: i64,
+    pub title: String,
+    pub embedding: Option<pgvector::Vector>,
+}
+
+#[derive(Debug, Deserialize)]
+pub struct CreatePost {
+    pub title: String,
+    pub content: String,
+}
+
+#[derive(Debug, Deserialize)]
+pub struct UpdatePost {
+    pub title: Option<String>,
+    pub content: Option<String>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct PostResponse {
+    pub id: i64,
+    pub title: String,
+    pub content: String,
+    pub created_at: DateTime<Utc>,
+    pub updated_at: DateTime<Utc>,
+    pub user_id: i64,
+    pub related_posts: Vec<Post>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct GraphNode {
+    pub id: String,
+    pub name: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct GraphLink {
+    pub source: String,
+    pub target: String,
+    pub value: f32,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct GraphData {
+    pub nodes: Vec<GraphNode>,
+    pub links: Vec<GraphLink>,
+}
