+++ axum/src/posts/utils.rs
@@ -1,0 +1,5 @@
+use axum::http::StatusCode;
+
+pub fn internal_error<E: std::fmt::Display>(e: E) -> (StatusCode, String) {
+    (StatusCode::INTERNAL_SERVER_ERROR, format!("에러: {}", e))
+}
